backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
beautifulsoup4==4.12.3
google-generativeai==0.7.2
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.134.0
APScheduler==3.10.4
gunicorn==22.0.0
lxml>=5.0.0